requirements.txt
deepface
matplotlib
numpy
opencv_contrib_python
opencv_python
pandas
Pillow
plotly
PyAutoGUI
streamlit
tensorflow
keras
tf-keras